+++ app/api/services/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
-import { authOptions } from '../auth/[...nextauth]/route'
+import { authOptions } from '../../../lib/auth'
 import { services, Service } from '../../../lib/data'
 
 export async function GET() {
